website/requirements.txt
gunicorn
Flask
facenet-pytorch
opencv-python-headless
boto3
google-cloud-datastore
keras